Music!
The game now has audio, but it is only music for now.

Also a note on the audio:
Normally you put an audio listener on the camera, so you can have directional sounds or quieter sounds. I put the audio listener on a new different object, however, because we can just play sounds as they are, because powering up or dying doesn't need to be directional.