Added Main menu (bad) and win reset
When the games over, it goes back to the main menu where the players can play again or quit.

It also starts on the main menu now.

I also had to delete the DestroyOnLoad on the game manager to get the retrying thing to work. It didn't break anything that I could tell. I did it because I was banging my head against the code trying to figure out how to not do that before I gave up.

I also fixed an issue where it could say the loser won if they finished laps after the winner.